moved layout to views
