+++ client/components/auth/signinbutton/password-reset/template.hbs
