+++ client/components/auth/signupbutton/accept-terms-input.hbs
