+++ client/components/auth/signupbutton/template.hbs
