+++ client/components/auth/social-login/template.hbs
