+++ client/components/navbar/index.js
@@ -1,28 +1,13 @@
 const View = require('ampersand-view')
-const templateHbs = require('./navbar.hbs')
 
-var AuthView = require('./auth')
+var AuthComponent = require('components/auth')
 
 module.exports = View.extend({
-  template: templateHbs,
-  subviews: { // append each subview to the items container
+  template: require('./navbar.hbs'),
+  subviews: {
     user: {
       selector: '[data-hook=items-container]',
-      constructor: AuthView
+      constructor: AuthComponent
     }
-  },
-  render: function () {
-    this.renderWithTemplate(this)
-
-    // $(window).scroll(function () {
-    //   var scroll = $(window).scrollTop()
-    //   if (scroll >= 100) {
-    //     $('.navbar-default').addClass('sticky')
-    //     $('.navbar-default').removeClass('topScroll')
-    //   } else {
-    //     $('.navbar-default').removeClass('sticky')
-    //     $('.navbar-default').addClass('topScroll')
-    //   }
-    // })
   }
 })
