--- client/index.tpl.html
@@ -1,9 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-
-  </head>
-  <body itemscope itemtype="http://schema.org/Product">
-    <div id="root"></div>
-  </body>
-</html>
