+++ client/views/layout/footer/footer.hbs
