+++ client/views/layout/templates/head.hbs
