--- templates/index.handlebars
@@ -1,7 +1,0 @@
-<!DOCTYPE html>
-<!-- NOT-ENV APPLICATION TEMPLATE -->
-<html>
-  <link href="/{{cssFileName}}" rel="stylesheet" type="text/css">
-  <script src="/js/misc.js"></script>
-  <script src="/{{jsFileName}}"></script>
-</html>
