+++ templates/index.html
@@ -1,0 +1,7 @@
+<!doctype html>
+<html>
+  <head></head>
+  <body>
+    <div id="root"></div>
+  </body>
+</html>
