--- templates/index.staging.handlebars
@@ -1,6 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <link href="/{{cssFileName}}" rel="stylesheet" type="text/css">
-  <script src="/js/misc-prod.js"></script>
-  <script src="/{{jsFileName}}"></script>
-</html>
